Rework auth and permissions
The auth middleware needed to be reimplemented to support the new use
of Auth0. Permissions middleware has been introduced to allow easy
checking of permissions before performing potentially sensitive actions.

Auth middleware tests have been scrapped. As I was implementing them I
found that I was focusing more on making the tests pass than actually
testing code. This topic will be revisited when I find a convenient
window of time to really sit and consider what needs to be tested and
how to refactor the implementation so that it can be easily tested.

As of right now the implementation is very straightforward, so the need
for testing in this regard is not high. But proper tests need to be put
in place before additional features beyond the scope of the prototype
are rolled out.