Add IDP token management and guilds endpoint
An identity provider token management system is in place now to retrieve
a users Discord token from Auth0 first, and then manage it offline in
the API system itself. When an IDP token expires and a request to the
API fails, the system will use the stored refresh token to fetch a new
token and retry the request. This has been abstracted away in a utility
wrapper.

Additionally, an endpoint has been introduced to allow users to view the
guilds that they are a part of. In the future a similar method will be
used by the system to ensure that they have elevated permissions in the
Discord server for which they are trying to make actions on behalf of.

Some minor changes can be found in other sections of the software.